Updated SUGCON sites to Content-SDK v2
