+++ headapps/Sugcon2024/.env.remote.example
@@ -19,16 +19,22 @@ SITECORE_EDGE_CONTEXT_ID=
 # Will be used as a fallback if separate SITECORE_EDGE_CONTEXT_ID value is not provided
 NEXT_PUBLIC_SITECORE_EDGE_CONTEXT_ID=
 
+# Optional: custom Sitecore Edge Platform hostname (hostname or full URL).
+NEXT_PUBLIC_SITECORE_EDGE_PLATFORM_HOSTNAME=
+
+# Optional: custom Experience Edge hostname for media URL rewriting (e.g. staging).
+SITECORE_EXPERIENCE_EDGE_HOSTNAME=
+
 # An optional Sitecore Personalize scope identifier.
 # This can be used to isolate personalization data when multiple XM Cloud Environments share a Personalize tenant.
 # This should match the PAGES_PERSONALIZE_SCOPE environment variable for your connected XM Cloud Environment.
 NEXT_PUBLIC_PERSONALIZE_SCOPE=
 
 # Timeout (ms) for Sitecore CDP requests to respond within. Default is 400.
-PERSONALIZE_MIDDLEWARE_CDP_TIMEOUT=
+PERSONALIZE_PROXY_CDP_TIMEOUT=
 
 # Timeout (ms) for Sitecore Experience Edge requests to respond within. Default is 400.
-PERSONALIZE_MIDDLEWARE_EDGE_TIMEOUT=
+PERSONALIZE_PROXY_EDGE_TIMEOUT=
 
 # Sitecore Content SDK npm packages utilize the debug module for debug logging.
 # https://www.npmjs.com/package/debug
@@ -42,3 +48,4 @@ PERSONALIZE_MIDDLEWARE_EDGE_TIMEOUT=
 # Client ID and Secret used for Design Library functionality
 SITECORE_AUTH_CLIENT_ID=
 SITECORE_AUTH_CLIENT_SECRET=
+SITECORE_RENDERINGHOST_NAME=
