+++ headapps/Sugcon2024/package.json
@@ -28,14 +28,15 @@
     "@fortawesome/free-regular-svg-icons": "^7.1.0",
     "@fortawesome/free-solid-svg-icons": "^7.1.0",
     "@fortawesome/react-fontawesome": "^3.1.1",
-    "@next/third-parties": "^15.5.9",
-    "@sitecore-cloudsdk/core": "^0.5.4",
-    "@sitecore-cloudsdk/events": "^0.5.4",
-    "@sitecore-content-sdk/nextjs": "^1.0.0",
+    "@next/third-parties": "^16.1.1",
+    "@sitecore-content-sdk/analytics-core": "^2.0.0",
+    "@sitecore-content-sdk/events": "^2.0.0",
+    "@sitecore-content-sdk/nextjs": "^2.0.0",
+    "@sitecore-content-sdk/personalize": "^2.0.0",
     "bootstrap": "^5.3.8",
     "cheerio": "^1.1.2",
     "font-awesome": "^4.7.0",
-    "next": "^15.5.9",
+    "next": "^16.1.1",
     "next-intl": "^4.3.5",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
@@ -46,15 +47,15 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
-    "@sitecore-content-sdk/cli": "^1.0.0",
+    "@sitecore-content-sdk/cli": "^2.0.0",
     "@tailwindcss/postcss": "^4",
-    "@types/node": "^22.15.14",
+    "@types/node": "^24.10.4",
     "@types/react": "^19.1.10",
     "@types/react-dom": "^19.1.7",
     "@types/react-slick": "^0.23.13",
     "cross-env": "^10.0.0",
     "eslint": "^9.33.0",
-    "eslint-config-next": "^15.5.9",
+    "eslint-config-next": "15.5.3",
     "npm-run-all2": "^8.0.4",
     "sass": "^1.97.1",
     "tailwindcss": "^4",
