+++ headapps/Sugcon2024/sitecore.config.ts.example
@@ -12,7 +12,9 @@ export default defineConfig({
         process.env.NEXT_PUBLIC_SITECORE_EDGE_CONTEXT_ID ||
         '',
       clientContextId: process.env.NEXT_PUBLIC_SITECORE_EDGE_CONTEXT_ID,
-      edgeUrl: process.env.SITECORE_EDGE_URL || process.env.NEXT_PUBLIC_SITECORE_EDGE_URL,
+      edgeUrl:
+        process.env.NEXT_PUBLIC_SITECORE_EDGE_PLATFORM_HOSTNAME ||
+        process.env.SITECORE_EDGE_PLATFORM_HOSTNAME,
     },
     local: {
       apiKey: process.env.NEXT_PUBLIC_SITECORE_API_KEY || '',
@@ -32,7 +34,7 @@ export default defineConfig({
   },
   personalize: {
     scope: process.env.NEXT_PUBLIC_PERSONALIZE_SCOPE,
-    edgeTimeout: parseInt(process.env.PERSONALIZE_MIDDLEWARE_EDGE_TIMEOUT!, 10),
-    cdpTimeout: parseInt(process.env.PERSONALIZE_MIDDLEWARE_EDGE_TIMEOUT!, 10),
+    edgeTimeout: parseInt(process.env.PERSONALIZE_PROXY_EDGE_TIMEOUT!, 10),
+    cdpTimeout: parseInt(process.env.PERSONALIZE_PROXY_CDP_TIMEOUT!, 10),
   },
 });
