+++ headapps/Sugcon2024/src/Bootstrap.tsx
@@ -1,7 +1,8 @@
 'use client';
 import { useEffect, JSX } from 'react';
-import { CloudSDK } from '@sitecore-cloudsdk/core/browser';
-import '@sitecore-cloudsdk/events/browser';
+import { initContentSdk } from '@sitecore-content-sdk/nextjs';
+import { eventsPlugin } from '@sitecore-content-sdk/events';
+import { analyticsBrowserAdapter, analyticsPlugin } from '@sitecore-content-sdk/analytics-core';
 import config from 'sitecore.config';
 
 const Bootstrap = ({
@@ -23,15 +24,23 @@ const Bootstrap = ({
     }
 
     if (config.api.edge?.clientContextId) {
-      CloudSDK({
-        sitecoreEdgeUrl: config.api.edge.edgeUrl,
-        sitecoreEdgeContextId: config.api.edge.clientContextId,
-        siteName: siteName || config.defaultSite,
-        enableBrowserCookie: true,
-        cookieDomain: window.location.hostname.replace(/^www\./, ''),
-      })
-        .addEvents()
-        .initialize();
+      initContentSdk({
+        config: {
+          contextId: config.api.edge.clientContextId,
+          edgeUrl: config.api.edge.edgeUrl,
+          siteName: siteName || config.defaultSite,
+        },
+        plugins: [
+          analyticsPlugin({
+            options: {
+              enableCookie: true,
+              cookieDomain: window.location.hostname.replace(/^www\./, ''),
+            },
+            adapter: analyticsBrowserAdapter(),
+          }),
+          eventsPlugin(),
+        ],
+      });
     } else {
       console.error('Client Edge API settings missing from configuration');
     }
