+++ headapps/Sugcon2024/src/byoc/index.tsx
@@ -2,7 +2,7 @@
 
 import React, { JSX } from 'react';
 import * as FEAAS from '@sitecore-feaas/clientside/react';
-import * as Events from '@sitecore-cloudsdk/events/browser';
+import * as Events from '@sitecore-content-sdk/events';
 import { LayoutServicePageState, SitecoreProviderReactContext } from '@sitecore-content-sdk/nextjs';
 import '@sitecore/components/context';
 import dynamic from 'next/dynamic';
