+++ headapps/Sugcon2024/src/components/basic-components/content-block/ContentBlock.tsx
@@ -22,4 +22,4 @@ const ContentBlock = ({ fields }: ContentBlockProps): JSX.Element => (
   </div>
 );
 
-export default withDatasourceCheck()<ContentBlockProps>(ContentBlock);
+export default ContentBlock;
