+++ headapps/Sugcon2024/src/components/content-sdk/CdpPageView.tsx
@@ -1,14 +1,13 @@
 'use client';
 import { useEffect, JSX } from 'react';
 import { CdpHelper, useSitecore } from '@sitecore-content-sdk/nextjs';
-import { pageView } from '@sitecore-cloudsdk/events/browser';
+import { pageView } from '@sitecore-content-sdk/events';
 import config from 'sitecore.config';
 
 /**
  * This is the CDP page view component.
- * It uses the Sitecore Cloud SDK to enable page view events on the client-side.
- * See Sitecore Cloud SDK documentation for details.
- * https://www.npmjs.com/package/@sitecore-cloudsdk/events
+ * See Sitecore Content SDK documentation for details.
+ * https://www.npmjs.com/package/@sitecore-content-sdk/events
  */
 const CdpPageView = (): JSX.Element => {
   const {
