+++ headapps/Sugcon2024/tsconfig.json
@@ -37,11 +37,11 @@
     "esModuleInterop": true,
     "strictFunctionTypes": false,
     "module": "esnext",
-    "moduleResolution": "node",
+    "moduleResolution": "bundler",
     "allowSyntheticDefaultImports": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve",
+    "jsx": "react-jsx",
     "incremental": true,
     "plugins": [
       {
